# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRISM Engine API",
    "description": "Multi-dimensional STEAM career guidance: psychometric + financial vectorisation, financial constraint solver, parent-student conflict index, market blending, SWOT and roadmaps.\n\nAll /api/v1 responses use the envelope {success, data, error, meta}. MOCK_MODE is currently **True**.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analysis/runs": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Create Run",
        "description": "Runs the full pipeline: vectorise -> fit -> financial solver -> conflict index -> market blend ->\nfinal score -> sensitivity. The run is immutable and reproducible (see `reproducibility`).",
        "operationId": "create_run_api_v1_analysis_runs_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRunRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnalysisRun_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_v1_analysis_runs_get",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Page_AnalysisRunSummary__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/runs/{run_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_analysis_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AnalysisRun_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/compare": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Compare",
        "operationId": "compare_api_v1_analysis_compare_get",
        "parameters": [
          {
            "name": "run_a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run A"
            }
          },
          {
            "name": "run_b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run B"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RunComparison_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/runs/{run_id}/what-if": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "What If",
        "operationId": "what_if_api_v1_analysis_runs__run_id__what_if_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_WhatIfResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/runs/{run_id}/conflict": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Conflict",
        "description": "Parents/educators get the full breakdown; students get a gentle summary (visibility=summary).",
        "operationId": "conflict_api_v1_analysis_runs__run_id__conflict_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ConflictReport_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/runs/{run_id}/swot": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Swot",
        "operationId": "swot_api_v1_analysis_runs__run_id__swot_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "career_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Career Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_SwotReport_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analysis/runs/{run_id}/roadmap": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Roadmap",
        "operationId": "roadmap_api_v1_analysis_runs__run_id__roadmap_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "career_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Career Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Roadmap_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisRun": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "family_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "examples": [
              "baseline",
              "what_if"
            ]
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "duration_ms": {
            "type": "number",
            "title": "Duration Ms"
          },
          "reproducibility": {
            "$ref": "#/components/schemas/Reproducibility"
          },
          "data_quality": {
            "$ref": "#/components/schemas/DataQuality"
          },
          "student_vector": {
            "additionalProperties": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Normalised to [0, 1]"
            },
            "type": "object",
            "title": "Student Vector"
          },
          "composite_scores": {
            "$ref": "#/components/schemas/CompositeScores"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Weights"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "buckets": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/BucketItem"
              },
              "type": "array"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Bucket"
            },
            "type": "object",
            "title": "Buckets"
          },
          "conflict": {
            "$ref": "#/components/schemas/ConflictReport"
          },
          "sensitivity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SensitivityReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Links",
            "description": "Relative URLs for swot/roadmap/what-if"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "student_id",
          "family_id",
          "kind",
          "created_at",
          "duration_ms",
          "reproducibility",
          "data_quality",
          "student_vector",
          "composite_scores",
          "weights",
          "recommendations",
          "buckets",
          "conflict",
          "sensitivity",
          "links"
        ],
        "title": "AnalysisRun"
      },
      "AnalysisRunRequest": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "options": {
            "$ref": "#/components/schemas/RunOptions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "student_id"
        ],
        "title": "AnalysisRunRequest"
      },
...
      "Envelope_AnalysisRun_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[AnalysisRun]"
      },
...
      "Envelope_ConflictReport_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConflictReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[ConflictReport]"
      },
...
      "Envelope_Page_AnalysisRunSummary__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page_AnalysisRunSummary_"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[Page[AnalysisRunSummary]]"
      },
...
      "Envelope_Roadmap_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Roadmap"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[Roadmap]"
      },
      "Envelope_RunComparison_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunComparison"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[RunComparison]"
      },
...
      "Envelope_SwotReport_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SwotReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[SwotReport]"
      },
...
      "Envelope_WhatIfResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WhatIfResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[WhatIfResult]"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "VALIDATION_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Meta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "v1/0.1.0"
          },
          "mock": {
            "type": "boolean",
            "title": "Mock",
            "description": "true when the payload is a MOCK_MODE fixture",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "Meta"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "student",
          "parent",
          "educator",
          "admin"
        ],
        "title": "Role"
      },
...
      "RunOptions": {
        "properties": {
          "top_k": {
            "type": "integer",
            "maximum": 40.0,
            "minimum": 3.0,
            "title": "Top K",
            "default": 10
          },
          "scoring_config_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoring Config Version",
            "description": "null = active config"
          },
          "include_sensitivity": {
            "type": "boolean",
            "title": "Include Sensitivity",
            "default": true
          },
          "candidate_career_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Career Ids",
            "description": "Restrict the candidate set"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RunOptions"
      },
...
      "WhatIfRequest": {
        "properties": {
          "overrides": {
            "$ref": "#/components/schemas/WhatIfOverrides"
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "examples": [
              "Take an education loan"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "overrides"
        ],
        "title": "WhatIfRequest"
      },
...
    }
  }
}